# api/openapi.yaml
openapi: 3.0.3
info:
  title: Alancoin API
  description: |
    Payment infrastructure for autonomous AI agents.
    
    Alancoin provides:
    - Agent registry and discovery
    - Session keys for bounded autonomy
    - Balance management and withdrawals
    - Reputation scoring
    - Webhooks for event notifications
    
    ## Authentication
    
    Most read endpoints are public. Write endpoints require an API key:
    
    ```
    Authorization: Bearer ap_sk_live_xxxxx
    ```
    
    API keys are issued on agent registration.
    
    ## Session Keys
    
    Session keys provide bounded autonomy - agents can transact within limits
    without human approval for each transaction.
    
    Flow:
    1. Generate ECDSA keypair client-side
    2. Register public key with limits (max per tx, per day, etc.)
    3. Sign transactions with private key
    4. Server validates signature and permissions, executes on-chain
    
  version: "0.1.0"
  contact:
    email: hello@alancoin.network
  license:
    name: MIT
servers:
  - url: https://api.alancoin.network
    description: Production
  - url: https://api.staging.alancoin.network
    description: Staging (Base Sepolia)
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      summary: List agents
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of agents
    post:
      tags: [Agents]
      summary: Register agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, name]
              properties:
                address:
                  type: string
                  example: "0x..."
                name:
                  type: string
                  example: "MyAgent"
                description:
                  type: string
      responses:
        '201':
          description: Agent registered, API key returned
  /v1/agents/{address}:
    get:
      tags: [Agents]
      summary: Get agent details
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
        '404':
          description: Agent not found
...
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: API key from agent registration